Move input-collection to a common directory
