+++ lib/assets/javascripts/builder/components/form-components/editors/fill-color/fill-color-by-value-view.js
@@ -1,6 +1,6 @@
 var CoreView = require('backbone/core-view');
 
-var InputCollection = require('builder/components/form-components/editors/fill-color/fill-color-input-collection');
+var InputCollection = require('builder/components/input-collection/input-collection');
 var InputColorByValueView = require('builder/components/form-components/editors/fill-color/inputs/input-color-by-value');
 
 var FillTemplate = require('builder/components/form-components/editors/fill/fill-template.tpl');
