+++ lib/assets/javascripts/builder/components/form-components/editors/stroke-color/stroke-color-view.js
@@ -2,7 +2,7 @@ var CoreView = require('backbone/core-view');
 var DialogModel = require('builder/components/dialog/dialog-model');
 var DialogView = require('builder/components/dialog/dialog-view');
 
-var InputCollection = require('builder/components/form-components/editors/fill-color/fill-color-input-collection');
+var InputCollection = require('builder/components/input-collection/input-collection');
 var InputStrokeColor = require('builder/components/form-components/editors/stroke-color/inputs/input-stroke-color');
 
 var FillTemplate = require('builder/components/form-components/editors/fill/fill-template.tpl');
