+++ lib/assets/javascripts/builder/components/input-collection/input-collection.js
